early: keep stdio files open
At the end of the 'early_console' function, the
file descriptor is closed unconditionally. This
'close' call closes the stdio files if the fd
returned by the 'open(dev/console)' call equals
with any of the STD{IN,OUT,ERR}_FILENO values.
When this happens, all subsequent accesses to
the stdio files will fail and early console
access won't work.

To avoid this, don't close the file descriptor if
that equals with any of the STD*_FILENO values.

Signed-off-by: Gabor Juhos <juhosg@openwrt.org>