service instance: Improve handling of watchdog config changes
When the ubus call "ubus call service watchdog" is used, it's expected
that the service might change its watchdog values during runtime when
it pets the watchdog.  The instance restarting behavior was setup though
to restart the instance if these changed.  This causes any service which
adjusts its watchdog to be restarted via "/etc/init.d/SERVICENAME start"
which is probably not what was intended.

Introduce a new variable, self_managed, if the service adjusted these
values at runtime so that a "start" doesn't cause the entire service to
restart.

Signed-off-by: Tim Nordell <tnordell@airgain.com>